ADD MERCURY HETEROGENEOUS CHEMISTRY
Adds heterogeneous chemistry for divalent (HgII) species. This currently
results in the production of particulate mercury (HgP) which does not undergo
any further transformation in this development. Update includes uptake to aerosols
and clouds (in the chem/module_mosaic_set_heterogeneous.F module).
Reaction rates are passed to KPP as photolysis rates (similar to heterogeneous
reactions of halogens).