ADD CL2 EMISSIONS FROM SNOW AND SEA ICE
Update includes a first description of Cl2 emissions from snow and
sea ice. Emissions for Cl2 are added to module_airsnow_emissions.F
and includes (1) photochemical release (based on SZA) from continental
snow and sea ice, and (2) recycling of HOCl and ClONO2 on continental
snow and sea ice. NOTE: This is subject to testing and updates.